requirements.txt
fastapi~=0.100.1
uvicorn~=0.23.2
sqlalchemy~=2.0.19
python-decouple
psycopg2-binary  # Driver para PostgreSQL
httpx          # Cliente HTTP necesario para algunas características de FastAPI
python-dotenv~=1.0.0
pydantic~=2.1.1
requests~=2.31.0
openpyxl~=3.1.2
python-multipart
pytz~=2023.3.post1
pyjasper
Jinja2
pandas
schedule
apscheduler
starlette~=0.27.0